requirements.txt
coverage==3.7.1
Fabric==1.10.1
nose==1.3.6
NoseGAE==0.5.6
WebTest==2.0.18
mock==1.0.1
pyyaml
pylint==1.4.3
Pillow==2.8.1
oauth2==1.5.211
httplib2==0.9
python-coveralls
python-dateutil==2.1
wtforms==0.6.3
gaenv==0.1.8.post0
GoogleAppEngineCloudStorageClient==1.9.22.1
